cleanup even more
